+++ src/Message/AbstractMessage.php
@@ -28,5 +28,5 @@ namespace BackQ\Message;
 
 abstract class AbstractMessage
 {
-    abstract public function getRecipientsNumber();
+
 }
