+++ src/Publisher/DynamoSQS.php
@@ -1,0 +1,7 @@
+<?php
+namespace BackQ\Publisher;
+
+final class DynamoSQS extends AbstractPublisher
+{
+    protected $queueName = 'scheduled_process';
+}
